add async execute method

+++ lib/commercetools-api/src/Models/Cart/CartAddCustomLineItemActionBuilder.php
@@ -39,7 +39,7 @@ final class CartAddCustomLineItemActionBuilder implements Builder
     private $name;
 
     /**
-     * @var ?int
+     * @var ?float
      */
     private $quantity;
 
@@ -80,7 +80,7 @@ final class CartAddCustomLineItemActionBuilder implements Builder
     }
 
     /**
-     * @return null|int
+     * @return null|float
      */
     public function getQuantity()
     {
@@ -142,7 +142,7 @@ final class CartAddCustomLineItemActionBuilder implements Builder
     /**
      * @return $this
      */
-    public function withQuantity(?int $quantity)
+    public function withQuantity(?float $quantity)
     {
         $this->quantity = $quantity;
