+++ lib/commercetools-api/src/Models/Cart/CartRemoveLineItemActionBuilder.php
@@ -28,7 +28,7 @@ final class CartRemoveLineItemActionBuilder implements Builder
     private $lineItemId;
 
     /**
-     * @var ?int
+     * @var ?float
      */
     private $quantity;
 
@@ -56,7 +56,7 @@ final class CartRemoveLineItemActionBuilder implements Builder
     }
 
     /**
-     * @return null|int
+     * @return null|float
      */
     public function getQuantity()
     {
@@ -100,7 +100,7 @@ final class CartRemoveLineItemActionBuilder implements Builder
     /**
      * @return $this
      */
-    public function withQuantity(?int $quantity)
+    public function withQuantity(?float $quantity)
     {
         $this->quantity = $quantity;
