+++ lib/commercetools-api/src/Models/Cart/ExternalTaxRateDraft.php
@@ -32,7 +32,7 @@ interface ExternalTaxRateDraft extends JsonObject
      * If <code>subRates</code> are specified
      * then the <code>amount</code> can be omitted or it must be the sum of the amounts of all <code>subRates</code>.</p>
      *
-     * @return null|int
+     * @return null|float
      */
     public function getAmount();
 
@@ -68,7 +68,7 @@ interface ExternalTaxRateDraft extends JsonObject
 
     public function setName(?string $name): void;
 
-    public function setAmount(?int $amount): void;
+    public function setAmount(?float $amount): void;
 
     public function setCountry(?string $country): void;
