+++ lib/commercetools-api/src/Models/Cart/ScoreShippingRateInputDraft.php
@@ -16,9 +16,9 @@ interface ScoreShippingRateInputDraft extends ShippingRateInputDraft
     public const FIELD_SCORE = 'score';
 
     /**
-     * @return null|int
+     * @return null|float
      */
     public function getScore();
 
-    public function setScore(?int $score): void;
+    public function setScore(?float $score): void;
 }
