+++ lib/commercetools-api/src/Models/Me/MyCartRemoveLineItemAction.php
@@ -28,7 +28,7 @@ interface MyCartRemoveLineItemAction extends MyCartUpdateAction
     public function getLineItemId();
 
     /**
-     * @return null|int
+     * @return null|float
      */
     public function getQuantity();
 
@@ -49,7 +49,7 @@ interface MyCartRemoveLineItemAction extends MyCartUpdateAction
 
     public function setLineItemId(?string $lineItemId): void;
 
-    public function setQuantity(?int $quantity): void;
+    public function setQuantity(?float $quantity): void;
 
     public function setExternalPrice(?Money $externalPrice): void;
