+++ lib/commercetools-api/src/Models/TaxCategory/TaxRateDraftBuilder.php
@@ -26,7 +26,7 @@ final class TaxRateDraftBuilder implements Builder
     private $name;
 
     /**
-     * @var ?int
+     * @var ?float
      */
     private $amount;
 
@@ -64,7 +64,7 @@ final class TaxRateDraftBuilder implements Builder
      * If <code>subRates</code> are specified
      * then the <code>amount</code> can be omitted or it must be the sum of the amounts of all <code>subRates</code>.</p>
      *
-     * @return null|int
+     * @return null|float
      */
     public function getAmount()
     {
@@ -124,7 +124,7 @@ final class TaxRateDraftBuilder implements Builder
     /**
      * @return $this
      */
-    public function withAmount(?int $amount)
+    public function withAmount(?float $amount)
     {
         $this->amount = $amount;
